chore:update search name SaveSearchModal component
