--- packages/app/src/RenameSearchModal.tsx
@@ -1,99 +1,0 @@
-import { FormEvent, useEffect, useState } from 'react';
-import { Form, Modal } from 'react-bootstrap';
-import { toast } from 'react-toastify';
-import { Button as MButton, Text } from '@mantine/core';
-
-import api from './api';
-import { genEnglishExplanation } from './queryv2';
-
-export default function RenameSearchModal({
-  onHide,
-  show,
-  searchName,
-  searchID,
-  searchQuery,
-  onRenameSuccess,
-}: {
-  onHide: () => void;
-  show: boolean;
-  searchName: string;
-  searchID: string;
-  searchQuery: string;
-  onRenameSuccess: (responseData: { _id: string }) => void;
-}) {
-  const updateLogView = api.useUpdateLogView();
-
-  const [parsedEnglishQuery, setParsedEnglishQuery] = useState<string>('');
-
-  useEffect(() => {
-    genEnglishExplanation(searchQuery).then(q => {
-      setParsedEnglishQuery(q);
-    });
-  }, [searchQuery]);
-
-  const onSubmitLogView = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    const target = e.target as typeof e.target & {
-      name: { value: string };
-    };
-    // TODO: fallback to parsedEnglishQuery for better UX ??
-    const name = target.name.value || parsedEnglishQuery;
-
-    updateLogView.mutate(
-      {
-        id: searchID,
-        name,
-        query: searchQuery,
-      },
-      {
-        onSuccess: response => {
-          onRenameSuccess(response.data);
-        },
-        onError: () => {
-          toast.error(
-            'An error occured. Please contact support for more details.',
-          );
-        },
-      },
-    );
-  };
-
-  return (
-    <Modal
-      aria-labelledby="contained-modal-title-vcenter"
-      centered
-      onHide={onHide}
-      show={show}
-      size="lg"
-    >
-      <Modal.Body className="bg-grey rounded">
-        <h5 className="text-muted">Rename Search</h5>
-        <Form onSubmit={onSubmitLogView}>
-          <Form.Group className="mb-3 mt-4">
-            <Text span fw="bold">
-              Query:
-            </Text>
-            <Text span> {searchQuery}</Text>
-          </Form.Group>
-          <Form.Group className="mb-2 mt-2">
-            <Form.Label className="text-start text-muted fs-7">Name</Form.Label>
-            <Form.Control
-              className="border-0 mb-4 px-3"
-              id="name"
-              name="name"
-              placeholder={parsedEnglishQuery}
-              size="sm"
-              type="text"
-              autoFocus
-              defaultValue={searchName}
-            />
-          </Form.Group>
-          <MButton size="sm" variant="light" type="submit">
-            Rename
-          </MButton>
-        </Form>
-      </Modal.Body>
-    </Modal>
-  );
-}
