+++ packages/app/src/SaveSearchModal.tsx
@@ -10,14 +10,23 @@ export default function SaveSearchModal({
   onHide,
   show,
   searchQuery,
+  mode,
+  searchName,
+  searchID,
   onSaveSuccess,
+  onUpdateSuccess,
 }: {
   onHide: () => void;
   show: boolean;
   searchQuery: string;
+  mode: 'save' | 'update';
+  searchName: string;
+  searchID: string;
   onSaveSuccess: (responseData: { _id: string }) => void;
+  onUpdateSuccess: (responseData: { _id: string }) => void;
 }) {
   const saveLogView = api.useSaveLogView();
+  const updateLogView = api.useUpdateLogView();
 
   const [parsedEnglishQuery, setParsedEnglishQuery] = useState<string>('');
 
@@ -36,22 +45,42 @@ export default function SaveSearchModal({
     // TODO: fallback to parsedEnglishQuery for better UX ??
     const name = target.name.value || parsedEnglishQuery;
 
-    saveLogView.mutate(
-      {
-        name,
-        query: searchQuery,
-      },
-      {
-        onSuccess: response => {
-          onSaveSuccess(response.data);
+    if (mode === 'update') {
+      updateLogView.mutate(
+        {
+          id: searchID,
+          name,
+          query: searchQuery,
         },
-        onError: () => {
-          toast.error(
-            'An error occurred. Please contact support for more details.',
-          );
+        {
+          onSuccess: response => {
+            onUpdateSuccess(response.data);
+          },
+          onError: () => {
+            toast.error(
+              'An error occured. Please contact support for more details.',
+            );
+          },
         },
-      },
-    );
+      );
+    } else {
+      saveLogView.mutate(
+        {
+          name,
+          query: searchQuery,
+        },
+        {
+          onSuccess: response => {
+            onSaveSuccess(response.data);
+          },
+          onError: () => {
+            toast.error(
+              'An error occured. Please contact support for more details.',
+            );
+          },
+        },
+      );
+    }
   };
 
   return (
@@ -81,10 +110,11 @@ export default function SaveSearchModal({
               size="sm"
               type="text"
               autoFocus
+              defaultValue={searchName}
             />
           </Form.Group>
           <MButton size="sm" variant="light" type="submit">
-            Save
+            {mode === 'update' ? 'Update' : 'Save'}
           </MButton>
         </Form>
       </Modal.Body>
