+++ packages/app/src/SearchPage.tsx
@@ -40,7 +40,6 @@ import { withAppNav } from './layout';
 import LogSidePanel from './LogSidePanel';
 import LogTable from './LogTable';
 import { MemoPatternTableWithSidePanel } from './PatternTableWithSidePanel';
-import RenameSearchModal from './RenameSearchModal';
 import SaveSearchModal from './SaveSearchModal';
 import SearchInput from './SearchInput';
 import SearchPageActionBar from './SearchPageActionBar';
@@ -462,12 +461,12 @@ function SearchPage() {
   );
 
   const [showSaveSearchModal, setShowSaveSearchModal] = useState(false);
-  const [renameSearchModalData, setRenameSearchModalData] = useState<{
-    show: boolean;
+  const [saveSearchModalData, setSaveSearchModalData] = useState<{
     searchID: string | null;
+    mode: 'update' | 'save';
   }>({
-    show: false,
     searchID: null,
+    mode: 'save',
   });
   const [configAlertModalShow, setConfigAlertModalShow] = useState(false);
   const deleteLogView = api.useDeleteLogView();
@@ -693,8 +692,11 @@ function SearchPage() {
       </Head>
       <SaveSearchModal
         show={showSaveSearchModal}
+        mode={saveSearchModalData.mode}
         onHide={() => setShowSaveSearchModal(false)}
         searchQuery={displayedSearchQuery}
+        searchName={selectedSavedSearch?.name ?? ''}
+        searchID={saveSearchModalData.searchID ?? ''}
         onSaveSuccess={responseData => {
           toast.success('Saved search created');
           router.push(
@@ -711,34 +713,12 @@ function SearchPage() {
           refetchLogViews();
           setShowSaveSearchModal(false);
         }}
-      />
-      <RenameSearchModal
-        show={renameSearchModalData.show}
-        searchName={selectedSavedSearch?.name ?? ''}
-        onHide={() =>
-          setRenameSearchModalData({
-            show: false,
-            searchID: null,
-          })
-        }
-        searchID={renameSearchModalData.searchID || ''}
-        searchQuery={displayedSearchQuery}
-        onRenameSuccess={responseData => {
+        onUpdateSuccess={responseData => {
           toast.success('Saved search renamed');
-          router.push(
-            `/search/${responseData._id}?${new URLSearchParams({
-              q: searchedQuery,
-              ...(isLive
-                ? {}
-                : {
-                    from: searchedTimeRange[0].getTime().toString(),
-                    to: searchedTimeRange[1].getTime().toString(),
-                  }),
-            }).toString()}`,
-          );
           refetchLogViews();
-          setRenameSearchModalData({
-            show: false,
+          setShowSaveSearchModal(false);
+          setSaveSearchModalData({
+            mode: 'save',
             searchID: null,
           });
         }}
@@ -816,8 +796,9 @@ function SearchPage() {
             onClickUpdateLogView={onClickUpdateLogView}
             selectedLogView={selectedSavedSearch}
             onClickRenameSearch={() => {
-              setRenameSearchModalData({
-                show: true,
+              setShowSaveSearchModal(true);
+              setSaveSearchModalData({
+                mode: 'update',
                 searchID: selectedSavedSearch?._id ?? null,
               });
             }}
